# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leverpay.io API Documentation",
        "description": "Crypto payment gateway",
        "contact": {
            "email": "abdilkura@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.leverpay.io",
            "description": "Test server"
        }
    ],
    "paths": {
...
        "/api/v1/merchant/get-merchant-profile": {
            "get": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Get merchant profile",
                "operationId": "get merchant profile details",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/merchant/update-merchant-profile": {
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Update merchant profile",
                "operationId": "Update merchant profile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "address",
                                    "business_name",
                                    "phone",
                                    "country_id",
                                    "state_id",
                                    "city_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "business_name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "country_id": {
                                        "enum": "[1]"
                                    },
                                    "state_id": {
                                        "enum": "[1]"
                                    },
                                    "city_id": {
                                        "enum": "[1]"
                                    },
                                    "passport": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/merchant/get-merchant-currencies}": {
            "get": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Get merchant's currencies",
                "operationId": "get merchant's currencies",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/merchant/add-currencies": {
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Add merchant curriencies",
                "operationId": "Add merchant curriencies",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "curriencies"
                                ],
                                "properties": {
                                    "curriencies": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/merchant/get-merchant-keys}": {
            "get": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Get merchant's merchant keys",
                "operationId": "get merchant's merchant keys",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/merchant/change-mode": {
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Change to mode from test mode to live mode or vise versa",
                "operationId": "Change test mode to live ",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mode"
                                ],
                                "properties": {
                                    "mode": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://api.leverpay.io/oauth/authorize",
                        "tokenUrl": "https://api.leverpay.io/oauth/token",
                        "refreshUrl": "https://api.leverpay.io/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}